Move scorebox function to separate folder and file
In addition to breaking out the scorebox print function, replaced
the all in the import of baseball functions and explicitly identified
each function and variable.
